remove some warnings.

+++ src/54_departure_boards/DeparturesTableInterfaceElement.cpp
@@ -155,7 +155,7 @@ namespace synthese
 					if (displayedList.size() > __NombrePages + 2)
 						__NombrePages = displayedList.size () - 2;
 				}
-				if (__Pages != UNKNOWN_VALUE && __NombrePages > __Pages)
+				if (__Pages != UNKNOWN_VALUE && __NombrePages > (size_t)__Pages)
 					__NombrePages = __Pages;
 			}
